+++ src/App.css
@@ -138,4 +138,18 @@ button {
   font-family: "Lilita One", cursive;
 }
 
-/* <iframe src="https://giphy.com/embed/l3vRnoppYtfEbemBO" width="480" height="480" frameBorder="0" class="giphy-embed" allowFullScreen></iframe><p><a href="https://giphy.com/gifs/starfield-pico8-l3vRnoppYtfEbemBO">via GIPHY</a></p> */
+.searchHistory {
+  font-size: 1.25rem;
+  font-weight: 700;
+  margin-bottom: 100px;
+  color: rgb(41, 50, 100);
+  text-transform: uppercase;
+}
+
+.searchHistory h3 {
+  text-decoration: underline;
+}
+
+.searchHistory div:hover {
+  cursor: pointer;
+}
